% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{AvRatio}
\alias{AvRatio}
\title{Mid point ratio}
\usage{
AvRatio(x)
}
\arguments{
\item{x}{A summarized dive profile (summarized by the Broken Stick algorithm)}
}
\value{
a numeric value
}
\description{
Function for calculating the avratio of a summarized dive.
}
\details{
It actually looks bugged. that's probably why It has not been properly working for dive selection, but it may affect other functions like.
Yes, it is used in in NDE local function. It should calculate the deviaion of the deepest point from the center of the dive
}
