% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{BGC}
\alias{BGC}
\title{BGC: back-ground-color}
\usage{
BGC(col = "bisque", alpha = 1)
}
\arguments{
\item{col}{name of the color  in any format supported by R.}

\item{alpha}{numeric, from 0 to 1. It sets the transparency of the color}
}
\value{
nothng, just paint the background
}
\description{
Change the background of the plot
}
\details{
This function takes the 'actual' ranges of the plotting region (R adds
4% extra to the range of the variables) and draw a rectangle with the desired color.
This function is intended to be called by panel.first argument inside plot function.
It should take any color in any format supported by R
}
