% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kalman.R
\name{DayDep}
\alias{DayDep}
\title{DayDep}
\usage{
DayDep(Data)
}
\arguments{
\item{Data}{A dataframe}
}
\value{
a data.frame
}
\description{
Get minimum and maximum days
}
\details{
gets mimimum and maximum days
}
\author{
Fer Arce
}
