% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kalman.R
\name{DayDrift}
\alias{DayDrift}
\title{eso}
\usage{
DayDrift(test, zetas = TRUE, zeta.crit = 0.99)
}
\arguments{
\item{test}{ni idea}

\item{zetas}{lol}

\item{zeta.crit}{el critero}
}
\value{
data frame
}
\description{
metrics del drift rate
}
\details{
ufff
}
\author{
fer
}
