% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kalman.R
\name{DayPlot}
\alias{DayPlot}
\title{DayPlot}
\usage{
DayPlot(tes, add = FALSE, col = c("blue", "dodgerblue"),
  points = TRUE, ...)
}
\arguments{
\item{tes}{DataSet}

\item{add}{Logical value, add if the Day estimated to be added to an existing plot}

\item{col}{Colour of the lines and points}

\item{points}{logical, whether points should be plotted or not above the lines}

\item{...}{Generic parameters from base::plot function}
}
\value{
a Fucking plot
}
\description{
un plot
}
\details{
no hase falta disir nada mas
}
\author{
Fernando Arce
}
