% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkFun.R
\name{drift}
\alias{drift}
\title{Sink Rate Modelling}
\usage{
drift(M0, V0, a, Vm = 1.11, p = 1/2, link = c("dragp", "drag0"))
}
\arguments{
\item{M0}{initial mass of animal}

\item{V0}{initial volume of animal}

\item{a}{a proportionality constant.}

\item{Vm}{the volume per unit mass of the accreted mass}

\item{p}{the power in the V/A relation for the dragp link}

\item{link}{the link function}
}
\value{
an object of class \code{family}.  Essential a gaussian
family with a link function that relates the animal's mass to
expected drift rate.
}
\description{
GLM family object to fit mass models to observed sink rate data.
}
\details{
This function provides a family object for modelling sink rates of
animals of varying condition.

The terminal or free settling velocity of a sinking object of
density \eqn{\rho}, volume \eqn{V} and cross-sectional area \eqn{A} is
\deqn{v = -\left ( \frac{2g V}{C_{d} A} (\rho-1) \right )^{1/2}}{%
v = -( 2 g V (\rho-1)/ (Cd A) )^(1/2)}

The volume \eqn{V} is modelled in terms of mass \eqn{M}, an
initial mass \eqn{M_{0}}{M0} and volume \eqn{V_{0}}{V0}, and the volume
\eqn{V_{m}}{Vm} per unit mass of the accreted mass
\deqn{V = V_{0} + V_{m}(M-M_{0})}{V = V0 + Vm (M-M0)}.

The expression for free settling velocity is rewritten as
\deqn{v = -a \left ( R(V/V_{0})(M/V-1) \right )^{1/2}}{%
v = -a (R(V/V0) (M/V-1) )^(1/2)}
where \eqn{R(V)= k V/A} and \eqn{R(1)=1} and \eqn{a} and \eqn{k}
are constants.

The family provides two links that relate the expected sink rate
to the mass of the animal.  The \code{"dragp"} link corresponds to
\eqn{R(z) = z^p}, and the \code{"drag0"} link corresponds to
\eqn{R(z) = 1}.

If the animal grows as a sphere as it gains condition, then area
scales as radius squared and volume as radius cubed, so \eqn{V/A
\propto V^{1/3}}{V/A = k V^(1/3)} and \eqn{p=1/3}.  If the animal
grows as a cyclinder of constant length as it gains condition,
then area scales with radius and volume as radius squared, so
\eqn{V/A \propto V^{1/2}}{V/A = k V^(1/2)} and \eqn{p=1/2}.  The
\eqn{p=0} case corresponds to an animal that grows only in one
dimension as gains condition.
}
