% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{driftFilter}
\alias{driftFilter}
\title{Filtering dives}
\usage{
driftFilter(fo)
}
\arguments{
\item{fo}{Object of class 'data.frame' generated by NewVarsVect local function}
}
\value{
An object of class 'data.frame' containing those dives accepted by the filter.
}
\description{
This function filters a pre-processed data.frame. consult the referenced paper for  specific references to the procedure
}
