% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ele.R
\docType{data}
\name{ele}
\alias{ele}
\title{elephant seal pre-processed dataset}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{D1}{D1}
  \item{D2}{D2}
  \item{D3}{D3}
  \item{D4}{D4}
  \item{T1}{T1}
  \item{T2}{T2}
  \item{T3}{T3}
  \item{T4}{T4}
  \item{DE_DATE}{DE_DATE}
  \item{DIVE_DUR}{DIVE_DUR}
  \item{MAX_DEP}{MAX_DEP}
  \item{SURF_DUR}{SURF_DUR}
  \item{lat}{lat}
  \item{lon}{lon}
  \item{ref}{ref}
...
}
}
\source{
\url{http://www.imos.org.au}
}
\usage{
ele
}
\description{
An example of a seal dataset as tipically provided by SMRU
tags. There are slight differences between datasets from SMRU
tags. Some variables may be lower or upper case in different data
sets.
}
\keyword{datasets}
