% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{formatDives}
\alias{formatDives}
\title{formatDives}
\usage{
formatDives(Data, format)
}
\arguments{
\item{Data}{an object of class daata.frame}
}
\value{
a data.frame containing a subset of the variables needed
    to keep processing the data.frame
}
\description{
Function that takes a subset of variables and do some recoding and
dive removal for filtering purposes
}
\details{
This functions keeps the variables needed to apply the subsequents
procedures of this package.  It will make the Date of class
\code{POSIXct} if supplied as a factor and will make the
\code{ref} (id of the tag) of type character. It also sorts the
rows by \code{ref} and by \code{Date}.  It also removes any
duplicated dive and those not reaching 100 meter depth and shorter
than 300 seconds.
}
