% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kalman.R
\name{Kalman}
\alias{Kalman}
\title{Function for applying a Kalman filter to the drift dives}
\usage{
Kalman(Data, update = 1e+05, n.iter = 1000, n.chains = 3,
  n.adapt = 500, list = FALSE)
}
\arguments{
\item{Data}{An object of class 'data.frame' that needs to include at least date and drift rate}

\item{update}{Burn.in length.}

\item{n.iter}{posterior drawn length.}

\item{n.chains}{number of chains to be drawn.}

\item{n.adapt}{adaption length (for jags performance)}

\item{list}{Logical, if the object is a list or not}
}
\value{
an object of class 'list' that includes: 1) the original dataset (plus some remaping of variables), the model (a jags object), the output etc...
}
\description{
Function for running a kalman filter to a subset of potential drift dives.
}
\details{
This functions is a wrap to a kalman filter implemented in jags (Just another Gibs sampler).
It is particularly slow as it will require long chains. Current set up is a burn in of 400.000 iterations.
There is a version that can paralelize the chains and take advantage of multicore processors
}
