% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.R
\name{makeTheGam}
\alias{makeTheGam}
\title{make The Gam}
\usage{
makeTheGam(Data, dates = NULL, zetas = 0.5)
}
\arguments{
\item{Data}{drifta}

\item{dates}{dates}

\item{zetas}{zetas}
}
\value{
a list with predicted values
}
\description{
function for fit the gam
}
\details{
function for fitting a gam with a custom link function

comming soon
}
\author{
Fer Arce
}
