% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{maxTime}
\alias{maxTime}
\title{maximum time}
\usage{
maxTime(x)
}
\arguments{
\item{x}{a given dive containing the times and depths of the
broken-stick algoritm.}
}
\value{
the time when the maximi depth is reached. A \code{vector}
    of class code{numeric} and length one.
}
\description{
Get the time-stamp at maximum depth
}
\details{
Get the time at which the maximum depth of a given dive is
reached.
}
