% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{maxTimes}
\alias{maxTimes}
\title{maximum time}
\usage{
maxTimes(x)
}
\arguments{
\item{x}{vector}
}
\value{
a vector of something
}
\description{
Get the time at which the maximum depth of a given dive is reached.
}
\details{
Get the time-stamp at maximum depth
}
