% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{mDepthR}
\alias{mDepthR}
\title{MdepthR}
\usage{
mDepthR(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a value
}
\description{
MdepthR
}
\details{
Internal function to calculate mDepthR. Shouldn't be called by the
end user
}
