% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{ModRes}
\alias{ModRes}
\title{Least square residuals}
\usage{
ModRes(x, res = 5)
}
\arguments{
\item{x}{a summarized dive}

\item{res}{wich residual is requested \cr
\itemize{
\item If {1,2,3,4} it will return the residual for the {1,2,3,4} inflection points
\item {5} all residuals pasted into a string separated by dots
}}
}
\value{
A numeric value if a single residual is requested, or a character string if all are requested
}
\description{
ModRes generates the residuals of a linear model by fitting the depth points (D1..D4) vs the time points (T1..T4). While the dimension may not be of importance, the sign of the residuals may be for some criteria. It only returns the residual values, not the full object of class 'lm'.
}
