% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nde.R
\name{NDE}
\alias{NDE}
\title{Drift Rate estimator}
\usage{
NDE(d, extract = "nde")
}
\arguments{
\item{d}{a single summarized dive row with variables generated by
calling function}

\item{extract}{flag for determining which output is requested to
the function \cr \itemize{ \item 'nde': call for retrieving
the drift rate estimate \item 'ds': call for retrieving the
drifting segment \item 'both': call for retrieving both values
}}
}
\value{
a numeric value, either the drift rate or the drift
    segment, or a vector containing both.
}
\description{
Function that calculates the drift rate and the drift segment of a
given dive.
}
\details{
The selection procedure is based on the 'order' of the generation
of the inflection points, plus a number of criteria, that helps to
determine if the dive is a drift dive, and, if so, whether is a
positive or negative drift, and which one is the drifting
segment. The drift rate is then calculated as the difference of
depths at the start and end of the drifting segment divided by its
duration.
}
