% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{newVarsVect}
\alias{newVarsVect}
\title{New variable generator}
\usage{
newVarsVect(Data = Data, t = FALSE)
}
\arguments{
\item{Data}{an object of class 'data.frame' containing at least the Broken-stick output ([T1...T4], [D1...D4]), dive duration (DIVE_DUR)}

\item{t}{\code{Logical}. wheter t1...t4 exist on the Data or not. If FALSE, they are computed}
}
\value{
this function returns the provided dataset with newly added columns named as variables described above, with the calculations.
}
\description{
Function for generating the variables required to apply the
filtering process. This function has been tested with tables
extracted from the Integrated Marine Observing System (IMOS,
Australia) freely available data, and it should work with any
Sea-mammal Research Unit (SMRU, St. Andrews, Scotland) provided
file. please contact the author if it does not work with your data
to find a custom adaptation suitable for your data.
}
\details{
Newly computed variables: \cr
\itemize{
\item descspeed: initial descending speed (from surface to the 1 infl point)
\item ascspeed: last ascending speed (from the 4 infl point to the surface)
\item daratio: the ratio between both speeds
\item max.depth: maximum depth (it's duplicated but both names are required)
\item min.depth: minimum depth of the inflection points
\item max.time: time spend before reaching max.depth
\item avratio: Deviation of the max.depth point from the middle of the inflected dive
\item hp1:  length of the first segment
\item hp2:  length of the second segment
\item hp3:  length of the third segment
\item Sa: for getting avratio
\item Sb: for getting avratio
\item d1: ratio between max.depth and the depth at the given inflection point (=1 if the given inflection point holds the max.depth). This calculates the first
\item d2: same ratio for the second inflection point
\item d3: same ratio for the third inflection point
\item d4: same ratio for the fourth inflection point
\item modres1: residual from a fitted linear model along the inflection points.Value for the first inflection point
\item modres2: for the second inflection point
\item modres3: for the third inflection point
\item modres4: for the fourth inflection point
\item propseg1: proportion of the dive along the first segment
\item propseg2: proportion of the dive along the second segment
\item propseg3: proportion of the dive along the third segment
}
}
