% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kalman.R
\name{parallelKalman}
\alias{parallelKalman}
\title{parallelised version of the Kalman filter.}
\usage{
parallelKalman(Data = Data)
}
\arguments{
\item{Data}{a data set to be kalmaned}
}
\value{
work in progress
}
\description{
Function to parallelise the Kalman filter
}
\details{
in progress. This function needs still a proper way to process the
output. It whould at some point be mixed as there are three kalman
instances. By using three cores, the coptutation time is reduced to half
}
