% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDrift}
\alias{plotDrift}
\title{plotDrift}
\usage{
plotDrift(Data, ID = FALSE, ...)
}
\arguments{
\item{Data}{}

\item{ID}{logical. If true it will print some axis and the tag id on the plot}
}
\description{
Function for making the vignette plot
}
\details{
This function wraps some code to make nice looking plots for the vignette.
}
\author{
Fer Arce
}
