% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kalman.R
\name{PostKalman}
\alias{PostKalman}
\title{Kalman post-processing}
\usage{
PostKalman(Data)
}
\arguments{
\item{Data}{an object of class 'kalman' generated by the local function Kalman (and or updateKalman)}
}
\value{
an object of class 'data.frame'
}
\description{
Function for post-processing kalman objects
}
\details{
This functions extrac the newly generated mu's and zetas and place them into the original Data frame, and return a simple data frame to keep processing
}
\author{
Fernando Arce
}
