% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kalman.R
\name{updateKalman}
\alias{updateKalman}
\title{updateKalman}
\usage{
updateKalman(data, update = 1e+05, n.iter = 1000, recompile = FALSE)
}
\arguments{
\item{data}{an object generated by the Kalman function}

\item{update}{number of iterations to be used for update the burn.in.}

\item{n.iter}{length of the drawn after the last burn.in.}

\item{recompile}{logical. if TRUE, the model will be recompiled}
}
\value{
an object of class 'kalman'
}
\description{
function for update objects generated by the local Kalman function
}
\details{
Requires an object generated by the local function 'Kalman'. If the object is loaded form a previously saved session, the model will need to be recompiled (and I will have to figure out soon how to do it)
}
\author{
Fernando Arce
}
