% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{avRatio}
\alias{avRatio}
\title{Mid point ratio}
\usage{
avRatio(x)
}
\arguments{
\item{x}{A summarized dive profile (summarized by the Broken Stick
algorithm)}
}
\value{
a \code{vector} of class \code{numeric} of length 1.
}
\description{
Function for calculating the avratio of a summarized dive.
}
\details{
This functions calculates the deviation of the deepest point from
the middle of the dive.
}
