% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{driftFilter}
\alias{driftFilter}
\title{Filtering dives}
\usage{
driftFilter(Data)
}
\arguments{
\item{Data}{Object of class \code{data.frame} generated by
\code{NewVarsVect}}
}
\value{
An object of class \code{data.frame} containing only the
    dives accepted by the filter.
}
\description{
This function filters a pre-processed data.frame. Consult the
paper for the specific criteria
}
\details{
This function removes all dives that does not match the criteria
proposed in Arce et al (2019) as characteristic of Drift dives.
}
