% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elek.R
\docType{data}
\name{elek}
\alias{elek}
\title{Post-processed seal}
\format{A list with 6 objects:
\describe{
  \item{Data}{Dataset imputed to the kalman filter}
  \item{model}{model definition}
  \item{kalman}{kalman expression}
  \item{mns}{summaries}
  \item{duration}{length of the run}
  \item{burn.in}{burn.in length, in iterations}
...
}}
\source{
\url{http://www.imos.org.au}
}
\usage{
elek
}
\description{
An example of a seal dataset as tipically provided by SMRU tags
}
\keyword{datasets}
