% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eleK.R
\docType{data}
\name{eleK}
\alias{eleK}
\title{Post-processed example (for saving computing time on the vignette)}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Date}{Drift rate}
  \item{NDE}{timestamp for each dive}
  \item{zetas}{probability of being inside the drifting trajectory}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
eleK
}
\description{
An example of a seal after being fully processed by the package
}
\keyword{datasets}
