% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{formatDives}
\alias{formatDives}
\title{formatDives}
\usage{
formatDives(Data)
}
\arguments{
\item{Data}{an object of class daata.frame}
}
\value{
a data.frame containing a subset of the variables needed to keep processing the data.frame
}
\description{
function that makes a subset of variables with some recoding and dive removal for filtering purposes
}
\details{
This functions keeps the variables needed to apply the subsequents procedures of this package.
It will make the date of class POSIXct if supplied as a factor and will make the ref (id of the tag) of type character. it also sort the rows by tag id (ref) and by date (Date).
It also removes dives shallower than 100 meter depth and shorter than 300 seconds.
}
