% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kalman.R
\name{kalman}
\alias{kalman}
\title{Function for applying a Kalman filter to the drift dives}
\usage{
kalman(
  Data,
  update = 4e+05,
  n.iter = 10000,
  n.chains = 3,
  n.adapt = 1000,
  parallel = FALSE
)
}
\arguments{
\item{Data}{An object of class 'data.frame' that needs to include
at least date and drift rate}

\item{update}{Burn.in length.}

\item{n.iter}{posterior drawn length.}

\item{n.chains}{number of chains to be drawn.}

\item{n.adapt}{adaption length (for jags performance)}

\item{parallel}{logical. if TRUE, it will run 3 chains in three
different cores of the local machine.}
}
\value{
an object of class 'list' that includes: 1) the original
    dataset (plus some remaping of variables), the model (a jags
    object), the output etc...
}
\description{
Function for running a kalman filter to a subset of potential
drift dives.
}
\details{
This functions is a wrap to a kalman filter implemented in jags
(Just another Gibs sampler).  It is a bit slow as it requires
require long chains. Current set up is a burn in of 400.000
iterations. The chains can be run in parallel. If parallel is set
to TRUE, it will run three chains, each one in a different core
}
