% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kalman.R
\name{MaxTime}
\alias{MaxTime}
\title{Maximum time?}
\usage{
MaxTime(seal)
}
\arguments{
\item{seal}{una puta foca en formato lista}
}
\value{
un numero
}
\description{
De esta si que no me acuerdo
}
\details{
hablame del mar marinero
}
\author{
fer
}
