% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDrift}
\alias{plotDrift}
\title{plotDrift}
\usage{
plotDrift(Data, ID = FALSE, ...)
}
\arguments{
\item{Data}{a seal dataset in \code{data.frame} format}

\item{ID}{\code{logical}. If true it will print axis and the tag id
on the plot}

\item{...}{aditional arguments to be pasted to the generic plot
function}
}
\value{
a plot of the drift rate time series.
}
\description{
Function for plotting Drift rate time series using pre-determined
graphical parameters
}
\details{
This function wraps some code to make nice looking plots for the
vignette.
}
