% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kalman.R
\name{postKalman}
\alias{postKalman}
\title{Kalman post-processing}
\usage{
postKalman(Data)
}
\arguments{
\item{Data}{an object of class 'kalman' generated by the local
function Kalman (and or updateKalman)}
}
\value{
an object of class \code{data.frame}
}
\description{
Function for post-processing kalman objects
}
\details{
This functions extrac the mus and zetas from the \code{kalman}
output and place them into the \code{data.frame} used by the
\code{kalman} function
}
