% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kalman.R
\name{updateKalman}
\alias{updateKalman}
\title{Update the kalman fit}
\usage{
updateKalman(data, update = 1e+05, n.iter = 1000, recompile = FALSE)
}
\arguments{
\item{data}{an object generated by the Kalman function}

\item{update}{number of iterations to be used for update the
burn.in.}

\item{n.iter}{length of the drawn after the last burn.in.}

\item{recompile}{\code{logical}. if TRUE, the model will be recompiled}
}
\value{
same object tipe as the one returned by \code{kalman}
}
\description{
function for update objects generated by the \code{kalman} function
}
\details{
Requires an object generated by the function \code{kalman}. If
the object is loaded from a previously saved session, the model
will need to be recompiled.
}
